--- src/app/[locale]/components/Footer.tsx
@@ -1,3 +1,0 @@
-export const Footer = () => {
-  return <div>Footer</div>
-}
